refactor: :recycle: Better submodules names (Fixes #14)
